added a flag to selectively enable shadow dom
Users can use the Shadow DOM instead of the original, through the `data-shadow-root` property of the base element.

`data-shadow-dom` attribute should be `open` or `closed`, or implicitly use `open` when it is empty.

When Shadow DOM is enabled, it can protect the host styles by isolating style changes that occur inside the Worker.
Uses of `<link>` for external stylesheets inside the Worker are also isolated, so they also need to be moved inside the base element.

Resolves #1093